+++ packages/hud/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @weasel-js/hud
 
+## 0.6.0
+
+### Patch Changes
+
+- @weasel-js/theme@0.6.0
+
 ## 0.5.1
 
 ### Patch Changes