+++ packages/weasel-js/CHANGELOG.md
@@ -1,5 +1,11 @@
 # weasel-js
 
+## 0.6.0
+
+### Patch Changes
+
+- @weasel-js/core@0.6.0
+
 ## 0.5.1
 
 ### Patch Changes